implement send eiamil for foget password

+++ controllers/coursesController.js
@@ -53,13 +53,17 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 exports.createCourse = asyncHandler(async (req, res, next) => {
 
     req.body.bootcamp = req.params.bootcampId
+    req.body.user = req.user._id
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId)
 
     if (!bootcamp) {
         return next(new ErrorResponse(`No Bootcamp With the Id of ${req.params.bootcampId}`, 404))
     }
-
+    //Make sure that the bootcampowner is the user that is logged in
+    if (bootcamp.user.toString() !== req.user._id.toString() && req.user.role !== "adming") {
+        return next(new ErrorResponse(`User ${req.user._id} is not authorized to create course!`))
+    }
     const newCourse = await Course.create(req.body)
     if (newCourse) {
 
@@ -82,6 +86,11 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
         return next(new ErrorResponse(`No Course With the Id of ${req.params.bootcampId}`, 404))
     }
 
+    //Make sure user is couese owner
+    //we chek the user id in the cous, with the id of the logged in user
+    if (course.user.toString() !== req.user._id.toString() && req.user.role !== "adming") {
+        return next(new ErrorResponse(`User ${req.user._id} is not authorized to update course ${course._id}!`))
+    }
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
         runValidators: true
@@ -103,9 +112,13 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
     let course = await Course.findById(req.params.id)
 
-
+   
     if (!course) {
-        return next(new ErrorResponse(`No Course With the Id of ${req.params.bootcampId}`, 404))
+        return next(new ErrorResponse(`No Course With the Id of ${req.params.id}`, 404))
+    }
+    //Make sure that the bootcampowner is the user that is logged in
+    if (course.user.toString() !== req.user._id.toString() && req.user.role !== "adming") {
+        return next(new ErrorResponse(`User ${req.user._id} is not authorized to delete course!`))
     }
 
     await Course.findByIdAndDelete(req.params.id)
